Change from using floats to flex for story__shape elements
The intention of this project was to use floats as a learning exercise
despite the shift in modern development to using flex and grid.

Float won't be used on this specific element because I was driving
myself crazy trying to figure out why lines were appearing around the
image during transitions-- turns out this is a bug with chrome on macs
when animating transitions on floats. Changed to using flex even
though I think it would look fine on other browsers which is a bummer
ultimately because shape-outside only works on floats, not flex and
I really liked the look of the text wrapping around the circle img.